+++ src/rSVDdpd.cpp
@@ -33,7 +33,8 @@
 //' of \eqn{L_2} norm, such estimation lacks efficiency. Application of density power
 //' divergence bridges the gap.
 //' \deqn{DPD(f|g) = \int f^{(1+\alpha)} - (1 + \frac{1}{\alpha}) \int f^{\alpha}g + \frac{1}{\alpha} \int g^{(1 + \alpha)} }
-//' The parameter \code{alpha} must be between 0 and 1, lower \code{alpha} means less robustness
+//' The parameter \code{alpha} should be between 0 and 1, if not, then a warning is shown.
+//' Lower \code{alpha} means less robustness
 //' but more efficiency in estimation, while higher \code{alpha} means high robustness but 
 //' less efficiency in estimation. The recommended value of \code{alpha} is 0.3.
 //' The function tries to obtain the best rank one approximation of a matrix by minimizing 
@@ -72,7 +73,7 @@ Rcpp::List rSVDdpd(arma::mat X, float alpha, int nd = NA_INTEGER,
     }
     
     if ((alpha < 0) || (alpha > 1)) {
-        Rcpp::stop("Robustness parameter alpha must be between 0 and 1. Recommended value is 0.3");
+        Rcpp::warning("Robustness parameter alpha should be between 0 and 1. Recommended value is 0.3");
     }
     
     if ((tol < 0) || (eps < 0)) {